+++ libpgmodeler_ui/src/modelwidget.h
@@ -63,7 +63,7 @@ class ModelWidget: public QWidget {
 		disable_render_smooth;
 
 		//! \brief Indicates if the minimum object opacity used when appliyng fade out to objects
-		static float min_object_opacity;
+		static double min_object_opacity;
 
 		/*! \brief Stores the model that generates the copy/cut operation. This model is updated
 		from the destination model whenever a past/cut operation is done. */
